predicter: renamed module from classifier
* Renames the INSPIRE classifier module for predictions to not clash
  with Invenio classifier module.
