+++ inspire/modules/predicter/config.py
@@ -20,7 +20,7 @@
 # granted to it by virtue of its status as an Intergovernmental Organization
 # or submit itself to any jurisdiction.
 
-"""Config for classifier INSPIRE module."""
+"""Config for predicter INSPIRE module."""
 
 import os
